+++ CHANGELOG.md
@@ -34,6 +34,11 @@
 
 - Support Python 3.12.
 
+### Changed
+
+- Paths are now `pathlib.Path` objects.
+- `directory.directories` is now a `platformdirs.PlatformDirs` object; as a consequence, the `directory.AppDirsPath` class does not exist anymore.
+
 ### Removed
 
 - Dropped Python 3.7 support.